Chore : 파티생성 페이지 관련 컴포넌트 통합
 NewPartyHeader와 NewParty컴포넌트도 한 개의 NewParty 컴포넌트에서 관리하기로 하였습니다.
